% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Code.R
\name{mc_power_simple_binom}
\alias{mc_power_simple_binom}
\title{Calculate simple power}
\usage{
mc_power_simple_binom(
  N = 10^5,
  p_0,
  p_1,
  alpha = 0.05,
  sample_size,
  alternative = c("greater", "less", "two_sided")
)
}
\arguments{
\item{N}{Integer. Number of simulated values (Simulations)}

\item{p_0}{Numeric. Null hypothesis}

\item{p_1}{Numeric. Well-defined alternative hypothesis}

\item{alpha}{Numeric. Critical value}

\item{sample_size}{Integer. Total trials in binomial (Sample Size per Simulation)}

\item{alternative}{Character. A vector of three different possible alternative}
}
\value{
the estimated type II error rate, power of the test and Monte Carlo standard error for both
}
\description{
This function calculates the type II error rate and power of the test
using Monte Carlo simulated binomial samples.
Notice : this function calculate type II error rate and power when alternative hypothesis is well-define(i.e. p = p_1)
if the alternative is not well-define, we have to compute the whole power curve !!!
}
\examples{
set.seed(1)
mc_power_simple_binom(N = 20000, p_0 = 0.5, p_1 = 0.6,
                     alpha = 0.05, sample_size = 100,
                     alternative = "greater")

}
