% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Code.R
\name{mc_pval_null}
\alias{mc_pval_null}
\title{Calculate null p_value under H_0}
\usage{
mc_pval_null(
  N = 10^5,
  p_0,
  sample_size,
  alternative = c("greater", "less", "two_sided")
)
}
\arguments{
\item{N}{Integer. Number of simulated values (Simulations). Default 10^5.}

\item{p_0}{Numeric. Null hypothesis probability.}

\item{sample_size}{Integer. Total trials in binomial (Sample Size per Simulation).}

\item{alternative}{Character. One of "greater", "less", or "two_sided".}
}
\value{
A numeric vector of length \code{N} containing the simulated null p-values.
}
\description{
This function calculates the p_value using Monte Carlo hypothesis testing
}
\examples{
set.seed(1)
pvals <- mc_pval_null(10000, p_0 = 0.5,
                       sample_size = 50,
                       alternative = "two_sided")
hist(pvals)

}
