% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Code.R
\name{mc_type1_binom}
\alias{mc_type1_binom}
\title{Calculate Type I Error Rate}
\usage{
mc_type1_binom(
  N = 10^5,
  p_0,
  sample_size,
  alpha = 0.05,
  alternative = c("greater", "less", "two_sided")
)
}
\arguments{
\item{N}{Integer. Number of simulated values (Simulations). Default 10^5.}

\item{p_0}{Numeric. Null hypothesis probability.}

\item{sample_size}{Integer. Total trials in binomial (Sample Size per Simulation).}

\item{alpha}{Numeric. Critical value. Default 0.05.}

\item{alternative}{Character. One of "greater", "less", or "two_sided".}
}
\value{
A list containing the estimated type I error rate and Monte Carlo standard error.
}
\description{
This function calculates the type I error rate using Monte Carlo simulated binomial samples.
}
