% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Code.R
\name{mc_power_curve_binom}
\alias{mc_power_curve_binom}
\title{Calculate power curve}
\usage{
mc_power_curve_binom(
  N = 10^5,
  p_0,
  alpha = 0.05,
  sample_size,
  alternative = c("greater", "less", "two_sided")
)
}
\arguments{
\item{N}{Integer. Number of simulated values (Simulations)}

\item{p_0}{Numeric. Null hypothesis}

\item{alpha}{Numeric. Critical value}

\item{sample_size}{Integer. Total trials in binomial (Sample Size per Simulation)}

\item{alternative}{Character. A vector of three different possible alternative}
}
\value{
the estimated type II error rate, power of the test and Monte Carlo standard error for both
}
\description{
This function calculates the type II error rate and power of the test
using Monte Carlo simulated binomial samples,
Notice : this function calculate power curve when alternative hypothesis is not well-define
if the alternative is well-define(i.e. p = p_1), use the function mc_power_simple_binom!!!
}
\examples{
set.seed(1)
pc <- mc_power_curve_binom(N = 5000, p_0 = 0.5, alpha = 0.05,
                          sample_size = 100, alternative = "two_sided")
head(pc)
}
