% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Code.R
\name{mc_pval_binom}
\alias{mc_pval_binom}
\title{Calculate p_value under a observed data}
\usage{
mc_pval_binom(
  N = 10^5,
  p_0,
  s_obs,
  sample_size,
  alternative = c("greater", "less", "two_sided")
)
}
\arguments{
\item{N}{Integer. Number of simulated values (Simulations). Default 10^5.}

\item{p_0}{Numeric. Null hypothesis probability.}

\item{s_obs}{Integer. Number of success you observed in one data set.}

\item{sample_size}{Integer. Total trials in binomial (Sample Size per Simulation).}

\item{alternative}{Character. One of "greater", "less", or "two_sided".}

\item{alpha}{Numeric. Critical value. Default 0.05.}
}
\value{
the p_value under s_obs and Monte Carlo standard error for p_value
}
\description{
This function calculates the p_value using Monte Carlo sampling.
}
\examples{
set.seed(1)
pvals <- mc_pval_null(10000, p_0 = 0.5,
                       sample_size = 50,
                       alternative = "two_sided")
hist(pvals)
}
